refactor PaletteColors component, separate each section with its own folder, remove console.log
