+++ src/components/sections/main/PaletteHeader.js
