+++ src/components/sections/sidebar/LeftBottomCard.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Slider from "./slider";
+import Slider from "../../slider";
 import {
   NumberInput,
   NumberInputField,
