+++ src/components/sections/sidebar/SidebarContainer.js
@@ -4,7 +4,7 @@ import LeftBottomCard from "./LeftBottomCard";
 
 function SidebarContainer() {
   return (
-    <Container gridArea={"sidebar"} p={4} h={"100%"} overflow={"scroll"}>
+    <Container gridArea={"sidebar"} p={4} h={"100%"} overflow={"scroll"} mt={4}>
       <LeftUpperCard />
       <LeftBottomCard />
     </Container>
