+++ src/utils/chipColorsDummyData.js
@@ -1,0 +1,12 @@
+const chipColors = [
+  { id: 1, color: "#054849" },
+  { id: 2, color: "#8ef08a" },
+  { id: 3, color: "#0081ec" },
+  { id: 4, color: "#f9c2f5" },
+  { id: 5, color: "#0ae2ff" },
+  { id: 6, color: "#f5be01" },
+  { id: 7, color: "#ff6c37" },
+  { id: 8, color: "#744012" },
+];
+
+export default chipColors;
